backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
pillow>=9.5.0
numpy>=1.20.0
python-dotenv>=0.20.0
requests>=2.28.0
openai>=1.0.0
google-generativeai>=0.1.0 